maint: Fix some coverity warnings (#259)
* Fix some coverity errors

may fix the following:

https://scan9.scan.coverity.com/#/project-view/60713/15801?selectedIssue=425120

* CWE-197: Numeric Truncation Error

https://scan9.scan.coverity.com/#/project-view/60713/15801?selectedIssue=425079

* CWE-457: Use of Uninitialized Variable

https://scan9.scan.coverity.com/#/project-view/60713/15801?selectedIssue=425076

* 425154 Uninitialized scalar variable

https://scan9.scan.coverity.com/#/project-view/60713/15801?selectedIssue=425154

+++ src/NetPanzer/Classes/Network/ConnectNetMessage.cpp
@@ -159,11 +159,11 @@ void ConnectMesgServerGameSettings::setTimeLimit(Sint32 timeLimit) {
 }
 
 time_t ConnectMesgServerGameSettings::getElapsedTime() const {
-  return (time_t)ltoh32((Uint32)elapsed_time);
+  return (time_t)ltoh32((Uint64)elapsed_time);
 }
 
 void ConnectMesgServerGameSettings::setElapsedTime(time_t elapsedTime) {
-  elapsed_time = (time_t)htol32((Uint32)elapsedTime);
+  elapsed_time = (time_t)htol32((Uint64)elapsedTime);
 }
 
 Sint32 ConnectMesgServerGameSettings::getFlagTime() const {
